% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsTable_exp.R
\name{eventsTable_exp}
\alias{eventsTable_exp}
\title{Computes the Expected Number of Events by Treatment Group and Biomarker Status}
\usage{
eventsTable_exp(dfx)
}
\arguments{
\item{dfx}{The data frame returned by join_parent()}
}
\value{
The expected number of events by treatment group and biomarker status
}
\description{
Computes the Expected Number of Events by Treatment Group and Biomarker Status
}
\examples{
library(survival)
library(dplyr)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select(.data$nrx, .data$survtime, .data$survstat) \%>\%
  arrange( .data$nrx, .data$survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survfit(kmdef ~ dfp$rx )
ppos  <- 0.30    # proportion of subjects expected to be M+ at the beginning of the study.
HR    <- c(1,2)  # biomarker hazard ratios in the M- and M+ subgroups.
dfm <- dissect_km(ppos=ppos, HR=HR, kmfit=kmfit)
dfm <- ppos_calc(dfm, ppos=ppos)
dfx <- join_parent(dfp=dfp, dfm=dfm)
eventsTable_exp(dfx)

}
