% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramsCheck.R
\name{paramsCheck}
\alias{paramsCheck}
\title{Pre-execution check of the user specified program parameters.}
\usage{
paramsCheck(ppos, HR, nsim, alpha, dfp)
}
\arguments{
\item{ppos}{Proportion of subjects expected to be M+ at time=0.}

\item{HR}{An array of length 2 containing the biomarker hazard ratio for the
standard treatment group and the experimental (targeted) treatment groups.}

\item{nsim}{The number of simulated trials to  be generated. Set to 0 if no simulations.}

\item{alpha}{The critical p-value for rejecting the null hypotheses that the ratio
of the biomarker (or treatment) hazard ratios = 1.}

\item{dfp}{A data frame with a row for each individual in the parent trial.
This data frame must include the variables: survtime, survstat, and nrx.
survtime (dbl) is the time at risk of the event of interest.
Survstat (dbl) an indicator for whether the subject evented (=1) or is censored (=0).
nrx (dbl) assigned treatment: (0=standard treatment, 1=experimental (targeted) treatment).}
}
\value{
NULL.
}
\description{
Pre-execution check of the user specified program parameters.
}
\examples{
data(parentTrial)
# Recodes to make data consistent with program requirements.
parentTrial$nrx  <- ifelse(parentTrial$nrx == 1, 0, 1)        # recode $rx (1,2) -> (0,1)
parentTrial$survstat <- ifelse((parentTrial$survstat == 1), 1, 0) # censor ind. 1=evt; 0=cens
paramsCheck(ppos=0.50, HR=c(1,2), nsim=1000, alpha=0.05, dfp=parentTrial)

}
