% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parentTrial.R
\docType{data}
\name{parentTrial}
\alias{parentTrial}
\title{Example Parent Trial data file used in PRBpower}
\format{
\subsection{\code{parentTrial}}{

A data frame with 500 rows (one row per study participant) and 5 columns.
}
}
\source{
These data were generated using random number generators.
}
\usage{
data(parentTrial)
}
\description{
This dataset is contrived and used only to demonstrate how to use of PRBpower.
}
\details{
\describe{
\item{nrx}{(num) treatment code 0=standard treatment, 1=experimental(targeted treatment)}
\item{survtime}{(num) Time at risk of the study event}
\item{survstat}{(num) an indicator for whether the inidividual evented; 0=censored, 1=evented }
}
}
\keyword{datasets}
