% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PRB.R
\name{plot_PRB}
\alias{plot_PRB}
\title{Commonly Created Plots for Prospective-Retrospective Biomarker Study Planning.}
\usage{
plot_PRB(dfx, plotType, ppos, HR)
}
\arguments{
\item{dfx}{The data frame returned by join_parent().}

\item{plotType}{1= Survival by treatment group.  2= Survival of the standard
treatment group and the expected survival of biomarker-negative subgroups
within the standard treatment group. 3= Survival of experimental (Targeted)
treatment group and expected survival of the biomarker-positive and
biomarker-negative subgroups within this treatment group.}

\item{ppos}{Expected proportion of the study population that are biomarker-positive
at time=0.}

\item{HR}{an array of length=2 with the biomarker hazard ratios for the standard
treatment group and the experimental treatment group.}
}
\value{
A ggplot2 figure object.
}
\description{
Commonly Created Plots for Prospective-Retrospective Biomarker Study Planning.
}
\examples{
library(survival)
library(dplyr)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select( .data$nrx, .data$survtime, .data$survstat) \%>\%
  arrange(.data$nrx, .data$survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survfit(kmdef ~ dfp$rx )
ppos  <- 0.30    # proportion of subjects expected to be M+ at the beginning of the study.
HR    <- c(1,2)  # biomarker hazard ratios in the M- and M+ subgroups.
dfm <- dissect_km(ppos=ppos, HR=HR, kmfit=kmfit)
dfm <- ppos_calc(dfm, ppos=ppos)
dfx <- join_parent(dfp=dfp, dfm=dfm)
plot_PRB(dfx=dfx, plotType=3, ppos=ppos, HR=HR)

}
