% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_PG.R
\name{power_PG}
\alias{power_PG}
\title{Calculates the Peterson-George Power of a Prospective-Retrospective Biomarker Study}
\usage{
power_PG(EE, HR, alpha)
}
\arguments{
\item{EE}{The expected number of events returned from eventsTable_exp().}

\item{HR}{An array of length 2 specifying the hypothesized biomarker hazard ratio
for each Treatment Group}

\item{alpha}{The critical p-value boundary for rejecting the null hypothesis which
is that the ratio of the biomarker hazard ratios (M+:M-) for individuals in the experimental
treatment group and the standard treatment group equals 1.}
}
\value{
(num) The calculated power
}
\description{
Calculates the Peterson-George Power of a Prospective-Retrospective Biomarker Study
}
\examples{
library(survival)
library(dplyr)
library(stats)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select(nrx, survtime, survstat) \%>\%
arrange(., nrx,survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survfit(kmdef ~ dfp$rx )
ppos  <- 0.30    # proportion of subjects expected to be M+ at the beginning of the study.
HR    <- c(1,2)  # biomarker hazard ratios in the M- and M+ subgroups.
dfm <- dissect_km(ppos=ppos, HR=HR, kmfit=kmfit)
dfm <- ppos_calc(dfm, ppos=ppos)
dfx <- join_parent(dfp=dfp, dfm=dfm)
evt.exp <- eventsTable_exp(dfx)
alpha <- 0.05
power_PG(EE=evt.exp, HR=HR, alpha=alpha)

}
