% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppos_calc.R
\name{ppos_calc}
\alias{ppos_calc}
\title{Computes the probability of being biomarker-positive (M+) at each event/censor
time in the parent study.}
\usage{
ppos_calc(dfm, ppos)
}
\arguments{
\item{dfm}{This object is returned from the function dissect_km().}

\item{ppos}{The proportion of subjects expected to be M+ at time=0 (baseline)}
}
\value{
A modified version of the input data frame, dfm, that includes an estimate
of being M+ at each event/censor time in the parent study.
}
\description{
Computes the probability of being biomarker-positive (M+) at each event/censor
time in the parent study.
}
\examples{
library(survival)
library(dplyr)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select(nrx, survtime, survstat) \%>\%
arrange(., nrx,survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- survival::Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survival::survfit(kmdef ~ dfp$rx )
ppos  <- 0.30    # proportion of subjects expected to be M+ at the beginning of the study.
HR    <- c(1,2)  # biomarker hazard ratios in the M- and M+ subgroups.
dfm <- dissect_km(ppos=ppos, HR=HR, kmfit=kmfit)
ppos_calc(dfm=dfm, ppos=ppos)

}
