% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_gen.R
\name{sim_gen}
\alias{sim_gen}
\title{Analyze the Parent Clincal Trial Data with Simulated Biomarker Status.}
\usage{
sim_gen(dfx, nsim)
}
\arguments{
\item{dfx}{Data Frame returned from join_parent().}

\item{nsim}{Number of simulated trials to be generated and analyzed.}
}
\value{
Data Frame summarizing the results from a proportional hazards model of each simulated trial.
}
\description{
Analyze the Parent Clincal Trial Data with Simulated Biomarker Status.
}
\details{
The probability that a trial subject is biomarker positive is a function of her time
at risk of experiencing the study event, whether she is censored or evented, and her study treatment.
The probability of being marker positive for all unique values of these three variables is
calculated in ppos_calc(). The appropriate probability is then linked to each person in the parent
trial with join_parent(). This function generates nsim trials in which each person's biomarker status
is randomly generated proportional to the expected probability of being biomarker positive.

Each simulated trial is then analyzed with a proportional hazards (PH) model which includes covariates for
patient's assigned treatment, simulated biomarker status, and an interaction term for these two covariates.

This function returns a data frame summarizing the results of each simulated trial.
}
\examples{
library(survival)
library(dplyr)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select(nrx, survtime, survstat) \%>\%
arrange(., nrx,survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survfit(kmdef ~ dfp$rx )
ppos  <- 0.30    # proportion of subjects expected to be M+ at the beginning of the study.
HR    <- c(1,2)  # biomarker hazard ratios in the M- and M+ subgroups.
dfm <- dissect_km(ppos=ppos, HR=HR, kmfit=kmfit)
dfm <- ppos_calc(dfm, ppos=ppos)
dfx <- join_parent(dfp=dfp, dfm=dfm)
nsim <- 100
sim_gen(dfx=dfx, nsim=nsim)

}
