% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dissect_km.R
\name{dissect_km}
\alias{dissect_km}
\title{Estimates the Survival For Biomarker positive (M+) and -negative (M-) subjects.}
\usage{
dissect_km(ppos, HR, kmfit)
}
\arguments{
\item{ppos}{The proportion of subjects assumed to be biomarker positive at time=0.}

\item{HR}{An array with length=2 containing the biomarker hazard ratios under the
alternative hypothesis for each treatment group in the parent trial.}

\item{kmfit}{The output from function survival::survfit comparing the treatment
groups in the parent trial.}
}
\value{
A data frame containing K-M survival estimates for M+ and M- subjects
in each treatment group in the parent trial.
}
\description{
Estimates the Survival For Biomarker positive (M+) and -negative (M-) subjects.
}
\details{
Using the Kaplan-Meier estimated survival from the parent study, the
posited biomarker hazard ratios and the expected prevalence of the biomarker in
in the study population, this function computes the probability of being biomarker
positive for each subject. If Sj+ and Sj- are the estimated survival for biomarker-
positive and biomarker-negative subject in the jth treatment group, then
Sj+ = Sj-^(HR(j)).  Also if p+ is the prevalence of the biomarker, then
Sj = p+*Sj+ + (1-p+)*Sj-, where Sj is the observed survival of the subjects
in the jth treatment group of the parent trial.
}
\examples{
library(survival)
library(dplyr)
# read data from the parent clinical trial
data(parentTrial)
dfp <- parentTrial \%>\% select(nrx, survtime, survstat) \%>\%
arrange(., nrx,survtime)
dfp$nrx  <- ifelse(dfp$nrx==1, 0, 1)          # recode df$rx (1,2) -> (0,1)
dfp$survstat <- ifelse((dfp$survstat == 1), 1, 0) # recode censor indicator. 1=event; 0=censor
dfp$rx   <- factor(dfp$nrx, levels=c(0,1), labels=c("std", "Exp"),ordered=TRUE)
kmdef <- Surv(dfp$survtime, dfp$survstat == 1)
kmfit <- survfit(kmdef ~ dfp$rx )
dissect_km(ppos=0.30, HR=c(1,2), kmfit)

}
